flaskk8s/requirements.txt
apispec==6.3.0
attrs==23.1.0
Babel==2.12.1
blinker==1.6.2
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
dnspython==2.4.2
email-validator==1.3.1
Flask==2.3.3
Flask-AppBuilder==4.3.6
Flask-Babel==2.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Limiter==3.5.0
Flask-Login==0.6.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
limits==3.6.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.26.1
mdurl==0.1.2
numpy==1.24.4
ordered-set==4.1.0
packaging==23.1
pandas==2.0.3
pkgutil_resolve_name==1.3.10
prison==0.2.1
psycopg2-binary==2.9.7
Pygments==2.16.1
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
rich==13.5.2
rpds-py==0.10.2
six==1.16.0
SQLAlchemy==1.4.49
SQLAlchemy-Utils==0.41.1
typing_extensions==4.7.1
tzdata==2023.3
Werkzeug==2.3.7
wrapt==1.15.0
WTForms==3.0.1
zipp==3.16.2
